requirements.txt
pymongo~=3.12.0
python-dotenv~=0.19.0
discord.py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
jsons~=1.5.0
coloredlogs~=15.0.1